+++ Runtime/DialogueRunner/DialogueRunner.cs
@@ -726,8 +726,19 @@ namespace Yarn.Unity
                         // Run the line and wait for it to finish
                         await view.RunLineAsync(localisedLine, token);
                     }
+                    catch (System.OperationCanceledException)
+                    {
+                        // The line presenter cancelled (rather than returning.)
+                        // This probably wasn't intended - they should clean up
+                        // and return null.
+                        Debug.LogWarning($"Dialogue presenter {view.name} threw an {nameof(System.OperationCanceledException)} when running its {nameof(DialoguePresenterBase.RunLineAsync)} method. Dialogue presenters should not throw this exception; instead, clean up any needed user-facing content, and return.", view);
+                    }
                     catch (System.Exception e)
                     {
+                        // If a dialogue presenter throws an exception, we need
+                        // to return, because the dialogue runner is waiting for
+                        // our task to complete. We'll log the exception so that
+                        // it's not lost, and exit here.
                         Debug.LogException(e, view);
                     }
                 }
@@ -798,13 +809,32 @@ namespace Yarn.Unity
                 {
                     return;
                 }
-                var result = await view.RunOptionsAsync(localisedOptions, optionCancellationSource.Token);
-                if (result != null)
+                try
                 {
-                    // We no longer need the other views, so tell them to stop
-                    // by cancelling the option selection.
-                    optionCancellationSource.Cancel();
-                    dialogueSelectionTCS.TrySetResult(result);
+                    var result = await view.RunOptionsAsync(localisedOptions, optionCancellationSource.Token);
+                    if (result != null)
+                    {
+                        // We no longer need the other views, so tell them to stop
+                        // by cancelling the option selection.
+                        optionCancellationSource.Cancel();
+                        dialogueSelectionTCS.TrySetResult(result);
+                    }
+                }
+                catch (System.OperationCanceledException)
+                {
+                    // The options presenter cancelled (rather than returning
+                    // null.) This probably wasn't intended - they should clean
+                    // up and return null.
+                    Debug.LogWarning($"Dialogue presenter {view.name} threw an {nameof(System.OperationCanceledException)} when running its {nameof(DialoguePresenterBase.RunOptionsAsync)} method. Dialogue presenters should not throw this exception; instead, clean up any needed user-facing content, and return null.", view);
+                }
+                catch (System.Exception ex)
+                {
+                    // If a dialogue presenter throws an exception, we still
+                    // need to return a value, because the dialogue runner is
+                    // waiting for our task to complete. We'll log the exception
+                    // so that it's not lost, and exit here.
+                    Debug.LogException(ex, view);
+                    return;
                 }
             }